requirements.txt
flask
line-bot-sdk
google-cloud-aiplatform
google-genai
langchain
requests